Move LdapEntryManager initialization to separate class
