+++ oxLdapSample/src/main/java/org/gluu/ldap/LdapSampleBatchJob.java
@@ -1,21 +1,23 @@
 package org.gluu.ldap;
 
-import com.unboundid.ldap.sdk.Filter;
-import com.unboundid.util.StaticUtils;
-import org.apache.log4j.Logger;
-import org.apache.logging.log4j.Level;
-import org.apache.logging.log4j.status.StatusLogger;
-import org.gluu.site.ldap.persistence.BatchOperation;
-import org.gluu.site.ldap.persistence.LdapEntryManager;
-import org.xdi.ldap.model.CustomAttribute;
-import org.xdi.ldap.model.SearchScope;
-import org.xdi.log.LoggingHelper;
-
 import java.text.ParseException;
 import java.util.Arrays;
 import java.util.Calendar;
 import java.util.Date;
 import java.util.List;
+
+import org.apache.log4j.Logger;
+import org.apache.logging.log4j.Level;
+import org.apache.logging.log4j.status.StatusLogger;
+import org.gluu.site.ldap.persistence.BatchOperation;
+import org.gluu.site.ldap.persistence.LdapEntryManager;
+import org.gluu.site.ldap.persistence.exception.EntryPersistenceException;
+import org.xdi.ldap.model.CustomAttribute;
+import org.xdi.ldap.model.SearchScope;
+import org.xdi.log.LoggingHelper;
+
+import com.unboundid.ldap.sdk.Filter;
+import com.unboundid.util.StaticUtils;
 
 /**
  * Created by eugeniuparvan on 1/12/17.
@@ -30,12 +32,19 @@ public class LdapSampleBatchJob {
     }
 
     public static void main(String[] args) {
-        LdapSample ldapSample = new LdapSample();
-        final LdapEntryManager ldapEntryManager = ldapSample.createLdapEntryManager();
+		// Prepare sample connection details
+		LdapSampleEntryManager ldapSampleEntryManager = new LdapSampleEntryManager();
+
+		// Create LDAP entry manager
+		final LdapEntryManager ldapEntryManager = ldapSampleEntryManager.createLdapEntryManager();
 
         BatchOperation<SimpleTokenLdap> tokenLdapBatchOperation = new BatchOperation<SimpleTokenLdap>(ldapEntryManager) {
-            @Override
+        	
+        	private int processedCount = 0;
+
+        	@Override
             protected List<SimpleTokenLdap> getChunkOrNull(int batchSize) {
+        		log.info("Processed: " + processedCount);
                 final Filter filter = Filter.createPresenceFilter("oxAuthExpiration");
                 return ldapEntryManager.findEntries("o=gluu", SimpleTokenLdap.class, filter, SearchScope.SUB, new String[]{"oxAuthExpiration"}, this, 0, batchSize, batchSize);
             }
@@ -43,9 +52,15 @@ public class LdapSampleBatchJob {
             @Override
             protected void performAction(List<SimpleTokenLdap> objects) {
                 for (SimpleTokenLdap simpleTokenLdap : objects) {
-                    CustomAttribute customAttribute = getUpdatedAttribute("oxAuthExpiration", simpleTokenLdap.getAttribute("oxAuthExpiration"));
-                    simpleTokenLdap.setCustomAttributes(Arrays.asList(new CustomAttribute[]{customAttribute}));
-                    ldapEntryManager.merge(simpleTokenLdap);
+                	try {
+	                    CustomAttribute customAttribute = getUpdatedAttribute("oxAuthExpiration", simpleTokenLdap.getAttribute("oxAuthExpiration"));
+	                    simpleTokenLdap.setCustomAttributes(Arrays.asList(new CustomAttribute[]{customAttribute}));
+	                    ldapEntryManager.merge(simpleTokenLdap);
+	                    processedCount++;
+    				} catch (EntryPersistenceException ex) {
+    					log.error("Failed to update entry", ex);
+    					
+    				}
                 }
             }
         };
@@ -53,8 +68,12 @@ public class LdapSampleBatchJob {
 
 
         BatchOperation<SimpleSession> sessionBatchOperation = new BatchOperation<SimpleSession>(ldapEntryManager) {
-            @Override
+        	
+        	private int processedCount = 0;
+
+        	@Override
             protected List<SimpleSession> getChunkOrNull(int batchSize) {
+        		log.info("Processed: " + processedCount);
                 final Filter filter = Filter.createPresenceFilter("oxLastAccessTime");
                 return ldapEntryManager.findEntries("o=gluu", SimpleSession.class, filter, SearchScope.SUB, new String[]{"oxLastAccessTime"}, this, 0, batchSize, batchSize);
             }
@@ -62,9 +81,15 @@ public class LdapSampleBatchJob {
             @Override
             protected void performAction(List<SimpleSession> objects) {
                 for (SimpleSession simpleSession : objects) {
-                    CustomAttribute customAttribute = getUpdatedAttribute("oxLastAccessTime", simpleSession.getAttribute("oxLastAccessTime"));
-                    simpleSession.setCustomAttributes(Arrays.asList(new CustomAttribute[]{customAttribute}));
-                    ldapEntryManager.merge(simpleSession);
+                	try { 
+	                    CustomAttribute customAttribute = getUpdatedAttribute("oxLastAccessTime", simpleSession.getAttribute("oxLastAccessTime"));
+	                    simpleSession.setCustomAttributes(Arrays.asList(new CustomAttribute[]{customAttribute}));
+	                    ldapEntryManager.merge(simpleSession);
+	                    processedCount++;
+					} catch (EntryPersistenceException ex) {
+						log.error("Failed to update entry", ex);
+						
+					}
                 }
             }
         };
